rearranged footer, added KHRP logo to footer and favicon
